+++ forms/src/misc/formModel/index.ts
@@ -1,3 +1,4 @@
+export * from './interfaces/formModelBuilder.interface';
 export * from './interfaces/modelDecoratorMetadata';
 export * from './interfaces/modelPropertyDecoratorMetadata';
 export * from './interfaces/validator.interface';