+++ forms/src/misc/formModel/interfaces/validator.interface.ts
@@ -1,6 +1,8 @@
 import {AsyncValidatorFn, ValidatorFn} from '@angular/forms';
 import {Dictionary} from '@jscrpt/common';
 
+import {FormModelBuilderDefaultArgs} from './formModelBuilder.interface';
+
 /**
  * Factory function used for creaging ValidatorFn
  */
@@ -10,7 +12,7 @@ export interface ValidatorFnFactoryFn<TArg extends Dictionary<any> = any>
      * Function that creates ValidatorFn
      * @param args - Arguments passed for create function from owning class
      */
-    (args: TArg): ValidatorFn;
+    (args: TArg&FormModelBuilderDefaultArgs): ValidatorFn;
 }
 
 /**
@@ -22,5 +24,5 @@ export interface AsyncValidatorFnFactoryFn<TArg extends Dictionary<any> = any>
      * Function that creates AsyncValidatorFn
      * @param args - Arguments passed for create function from owning class
      */
-    (args: TArg): AsyncValidatorFn;
+    (args: TArg&FormModelBuilderDefaultArgs): AsyncValidatorFn;
 }