+++ forms/src/modules/formPipes/pipes/getControl/getControl.pipe.ts
@@ -3,18 +3,18 @@ import {AbstractControl, FormArray, FormGroup} from '@angular/forms';
 import {isBlank} from '@jscrpt/common';
 
 /**
- * Gets control from `FormGroup`
+ * Gets control from `FormGroup` or `FormArray`
  */
 @Pipe({name: 'getControl'})
-export class GetControlPipe<TFormGroup> implements PipeTransform
+export class GetControlPipe implements PipeTransform
 {
     /**
-     * Gets control from `FormGroup`
+     * Gets control from `FormGroup` or `FormArray`
      * @param value - FormGroup which contains requested control
      * @param control - Control name to be obtained
      * @param _valueHash - Parameter that can be used to change value, indicates that pipe input value has changed
      */
-    public transform(value: AbstractControl, control: keyof TFormGroup, _valueHash?: any): AbstractControl|null
+    public transform<TModel, TKey extends keyof TModel>(value: AbstractControl<TModel>, control: TKey, _valueHash?: unknown): AbstractControl<TModel[TKey]>|null
     {
         if(isBlank(value))
         {